Fixed SET FEATURE and CLEAR FEATURE control requests directed at an unconfigured endpoint causing request timeouts.

+++ LUFA/Drivers/USB/HighLevel/Template/Template_Endpoint_Control_W.c
@@ -24,7 +24,7 @@ uint8_t TEMPLATE_FUNC_NAME (const void* Buffer,
 		  
 		if (Endpoint_IsINReady())
 		{
-			uint8_t BytesInEndpoint = Endpoint_BytesInEndpoint();
+			uint16_t BytesInEndpoint = Endpoint_BytesInEndpoint();
 		
 			while (Length && (BytesInEndpoint < USB_ControlEndpointSize))
 			{
